### 2018-10-06 #999 @whedon
[PRE REVIEW]: Pingouin: statistics in Python
**Submitting author:** @raphaelvallat (<a href="http://orcid.org/0000-0003-1779-7653">Raphael Vallat</a>)
**Repository:** <a href="https://github.com/raphaelvallat/pingouin" target ="_blank">https://github.com/raphaelvallat/pingouin</a>
**Version:** v0.1.10
