### 2018-10-06 ce432ec
Added circ_vtest
