### 2018-10-06 fe60f13
Minor performance improvements
