+++ pingouin/pairwise.py
@@ -16,11 +16,11 @@ def _append_stats_dataframe(stats, x, y, xlabel, ylabel, effects, alpha,
     stats = stats.append({
         'A': xlabel,
         'B': ylabel,
-        'mean(A)': np.round(np.mean(x), 3),
-        'mean(B)': np.round(np.mean(y), 3),
+        'mean(A)': np.round(x.mean(), 3),
+        'mean(B)': np.round(y.mean(), 3),
         # Use ddof=1 for unibiased estimator (pandas default)
-        'std(A)': np.round(np.std(x, ddof=1), 3),
-        'std(B)': np.round(np.std(y, ddof=1), 3),
+        'std(A)': np.round(x.std(ddof=1), 3),
+        'std(B)': np.round(y.std(ddof=1), 3),
         'Type': effects,
         'Paired': paired,
         'tail': df_ttest.loc['T-test', 'tail'],
