### 2018-10-08 8ebc9a1
Added Henze Zirkler multivariate normality test
