+++ pingouin/parametric.py
@@ -5,9 +5,10 @@ import pandas as pd
 from pingouin import (_check_dataframe, _remove_rm_na, _remove_na,
                       _export_table, bayesfactor_ttest)
 
-__all__ = ["gzscore", "test_normality", "test_homoscedasticity", "test_dist",
-           "epsilon", "test_sphericity", "ttest", "rm_anova", "rm_anova2",
-           "anova", "anova2", "mixed_anova", "ancova", "ancovan"]
+__all__ = ["gzscore", "test_normality", "multivariate_normality",
+           "test_homoscedasticity", "test_dist", "epsilon", "test_sphericity",
+           "ttest", "rm_anova", "rm_anova2", "anova", "anova2", "mixed_anova",
+           "ancova", "ancovan"]
 
 
 def gzscore(x):
@@ -119,6 +120,122 @@ def test_normality(*args, alpha=.05):
         p = float(p)
 
     return normal, np.round(p, 3)
+
+
+def multivariate_normality(X, alpha=.05):
+    """Henze-Zirkler's Multivariate Normality Test.
+
+    Parameters
+    ----------
+    X : np.array
+        Data matrix of shape (n, p) where n are the observations and p the
+        variables.
+    alpha : float
+        Significance level.
+
+    Returns
+    -------
+    normal : boolean
+        True if X comes from a multivariate normal distribution.
+    p : float
+        P-value.
+
+    See Also
+    --------
+    test_normality : Test the univariate normality of one or more variables.
+    test_homoscedasticity : Test equality of variance.
+    test_sphericity : Mauchly's test for sphericity.
+
+    Notes
+    -----
+    The Henze-Zirkler test has a good overall power against alternatives
+    to normality and is feasable for any dimension and any smaple size.
+
+    Translated from the Matlab Robust Correlation toolbox.
+
+    References
+    ----------
+    .. [1] Henze, N., & Zirkler, B. (1990). A class of invariant consistent
+       tests for multivariate normality. Communications in Statistics-Theory
+       and Methods, 19(10), 3595-3617.
+
+    .. [2] Pernet, C. R., Wilcox, R. R., & Rousselet, G. A. (2013). Robust
+       correlation analyses: false positive and power validation using a new
+       open source Matlab toolbox. Frontiers in psychology, 3, 606.
+
+    Examples
+    --------
+    1. Test for multivariate normality of 2 variables
+
+        >>> import numpy as np
+        >>> from pingouin import multivariate_normality
+        >>> np.random.seed(123)
+        >>> mean, cov, n = [4, 6], [[1, .5], [.5, 1]], 30
+        >>> X = np.random.multivariate_normal(mean, cov, n)
+        >>> normal, p = multivariate_normality(X, alpha=.05)
+        >>> print(normal, p)
+            True 0.7523511059223078
+
+    2. Test for multivariate normality of 3 variables
+
+        >>> import numpy as np
+        >>> from pingouin import multivariate_normality
+        >>> np.random.seed(123)
+        >>> mean, cov = [4, 6, 5], [[1, .5, .2], [.5, 1, .1], [.2, .1, 1]]
+        >>> X = np.random.multivariate_normal(mean, cov, 50)
+        >>> normal, p = multivariate_normality(X, alpha=.05)
+        >>> print(normal, p)
[…]
