### 2018-10-08 #999 @cMadan
Hi @raphaelvallat! I'm happy to be editor for this submission. It looks like your submission is generally in order, though you can remove the `Citations to entries in paper.bib should be in rMarkdown format.` from the paper.

Do you have any suggestions for reviewers for this submission?
