### 2018-10-08 ebe2f20
removed citation rmarkdown from JOSS paper
