### 2018-10-08 b96d3da
BIG MOVE: moved and renamed distribution test
