+++ pingouin/tests/test_parametric.py
@@ -2,10 +2,8 @@ import pandas as pd
 import numpy as np
 
 from pingouin.tests._tests_pingouin import _TestPingouin
-from pingouin.parametric import (gzscore, test_normality, ttest, anova, anova2,
-                                 rm_anova, mixed_anova, test_dist, epsilon,
-                                 rm_anova2, ancova, ancovan,
-                                 multivariate_normality)
+from pingouin.parametric import (ttest, anova, anova2, rm_anova, mixed_anova,
+                                 rm_anova2, ancova, ancovan)
 from pingouin.datasets import read_dataset
 
 # Generate random data for ANOVA
@@ -31,64 +29,10 @@ df_nan.iloc[[4, 15], 0] = np.nan
 np.random.seed(1234)
 x = np.random.normal(scale=1., size=100)
 y = np.random.normal(scale=0.8, size=100)
-z = np.random.normal(scale=0.9, size=100)
 
 
 class TestParametric(_TestPingouin):
     """Test parametric.py."""
-
-    def test_gzscore(self):
-        """Test function gzscore."""
-        raw = np.random.lognormal(size=100)
-        gzscore(raw)
-
-    def test_test_normality(self):
-        """Test function test_normality."""
-        test_normality(x, alpha=.05)
-        test_normality(x, y, alpha=.05)
-
-    def test_multivariate_normality(self):
-        """Test function multivariate_normality."""
-        np.random.seed(123)
-        # With 2 variables
-        mean, cov, n = [4, 6], [[1, .5], [.5, 1]], 30
-        X = np.random.multivariate_normal(mean, cov, n)
-        normal, p = multivariate_normality(X, alpha=.05)
-        # Compare with the Matlab Robust Corr toolbox
-        assert normal
-        assert np.round(p, 3) == 0.752
-        # With 3 variables
-        mean, cov = [4, 6, 5], [[1., .5, .2], [.5, 1., .1], [.2, .1, 1.]]
-        X = np.random.multivariate_normal(mean, cov, 50)
-        normal, p = multivariate_normality(X, alpha=.01)
-
-    # def test_test_homoscedasticity(self):
-    #     """Test function test_homoscedasticity."""
-    #     test_homoscedasticity(x, y, alpha=.05)
-    #
-
-    def test_epsilon(self):
-        """Test function epsilon."""
-        df_pivot = df.pivot(index='Subject', columns='Time',
-                            values='Scores').reset_index(drop=True)
-        eps_gg = epsilon(df_pivot)
-        eps_hf = epsilon(df_pivot, correction='hf')
-        # Compare with ezANOVA
-        assert np.allclose([eps_gg, eps_hf], [0.9987509, 1])
-
-    # def test_test_sphericity(self):
-    #     """Test function test_sphericity."""
-    #     df_pivot = df.pivot(index='Subject', columns='Time',
-    #                         values='Scores').reset_index(drop=True)
-    #     _, W, _, _, p = test_sphericity(df_pivot, method='mauchly')
-    #     # Compare with ezANOVA
-    #     assert np.allclose([W, p], [0.9987493, 0.9643574])
-    #     # JNS (default method)
-    #     test_sphericity(df_pivot, method='jns')
-
-    def test_test_dist(self):
-        """Test function test_dist."""
-        test_dist(x)
 
     def test_ttest(self):
         """Test function ttest"""
