+++ pingouin/distribution.py
@@ -1,0 +1,540 @@
+import numpy as np
+import pandas as pd
+
+__all__ = ["gzscore", "normality", "multivariate_normality",
+           "homoscedasticity", "anderson", "epsilon", "sphericity"]
+
+
+def gzscore(x):
+    """Geometric standard (Z) score.
+
+    Parameters
+    ----------
+    x : array_like
+        Array of raw values
+
+    Returns
+    -------
+    gzscore : array_like
+        Array of geometric z-scores (same shape as x)
+
+    Notes
+    -----
+    Geometric Z-scores are better measures of dispersion than arithmetic
+    z-scores when the sample data come from a log-normally distributed
+    population.
+
+    Given the raw scores :math:`x`, the geometric mean :math:`\mu_g` and
+    the geometric standard deviation :math:`\sigma_g`,
+    the standard score is given by the formula:
+
+    .. math:: z = \dfrac{log(x) - log(\mu_g)}{log(\sigma_g)}
+
+    References
+    ----------
+
+    .. [1] https://en.wikipedia.org/wiki/Geometric_standard_deviation
+
+    Examples
+    --------
+    Standardize a log-normal array
+
+        >>> import numpy as np
+        >>> from pingouin import gzscore
+        >>> np.random.seed(123)
+        >>> raw = np.random.lognormal(size=100)
+        >>> print(raw.mean().round(3), raw.std().round(3))
+            1.849 2.282
+        >>> z = gzscore(raw)
+        >>> print(z.mean(), z.std())
+            0 0.995
+    """
+    from scipy.stats import gmean
+    # Geometric mean
+    geo_mean = gmean(x)
+    # Geometric standard deviation
+    gstd = np.exp(np.sqrt(np.sum((np.log(x / geo_mean))**2) / (len(x) - 1)))
+    # Geometric z-score
+    return np.log(x / geo_mean) / np.log(gstd)
+
+
+def normality(*args, alpha=.05):
+    """Shapiro-Wilk univariate normality test.
+
+    Parameters
+    ----------
+    sample1, sample2,... : array_like
+        Array of sample data. May be different lengths.
+
+    Returns
+    -------
+    normal : boolean
+        True if x comes from a normal distribution.
+    p : float
+        P-value.
+
+    See Also
+    --------
+    homoscedasticity : Test equality of variance.
+    sphericity : Mauchly's test for sphericity.
+
+    Examples
+    --------
+    1. Test the normality of one array.
+
+        >>> import numpy as np
+        >>> from pingouin import normality
+        >>> np.random.seed(123)
+        >>> x = np.random.normal(size=100)
+        >>> normal, p = normality(x, alpha=.05)
+        >>> print(normal, p)
+        True 0.27
+
+    2. Test the normality of two arrays.
+
+        >>> import numpy as np
+        >>> from pingouin import normality
+        >>> np.random.seed(123)
+        >>> x = np.random.normal(size=100)
+        >>> y = np.random.rand(100)
+        >>> normal, p = normality(x, y, alpha=.05)
+        >>> print(normal, p)
+        [True   False] [0.27   0.0005]
+    """
+    from scipy.stats import shapiro
+    k = len(args)
+    p = np.zeros(k)
+    normal = np.zeros(k, 'bool')
+    for j in range(k):
+        _, p[j] = shapiro(args[j])
+        normal[j] = True if p[j] > alpha else False
+
+    if k == 1:
+        normal = bool(normal)
+        p = float(p)
+
[…]
