+++ docs/generated/pingouin.test_homoscedasticity.rst
@@ -1,6 +1,0 @@
-pingouin.test\_homoscedasticity
-===============================
-
-.. currentmodule:: pingouin
-
-.. autofunction:: test_homoscedasticity
