+++ README.rst
@@ -166,9 +166,9 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 
 .. code-block:: python
 
-   from pingouin import test_normality
+   from pingouin import normality
    # Return a boolean (true if normal) and the associated p-value
-   test_normality(x, y)
+   normality(x, y)
 
 .. parsed-literal::
 
