+++ docs/generated/pingouin.homoscedasticity.rst
@@ -1,0 +1,6 @@
+pingouin.homoscedasticity
+=========================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: homoscedasticity
