### 2018-10-08 #999 @raphaelvallat
Hi @cMadan! Thank you for accepting to be the editor!

I just re-generated the PDF without the "Citation" part. I think that these reviewers would be a great fit:
- snastase
- nhejazi
- kellieotto
