### 2018-10-09 e1b94b6
Bugfix: renamed csv in manifest
