### 2018-10-09 40caabf
Updated todo and readme
