+++ pingouin/__init__.py
@@ -11,4 +11,4 @@ from .circular import *
 from .pairwise import *
 from .power import *
 
-__version__ = "0.1.9"
+__version__ = "0.1.10"
