+++ docs/build/html/_sources/index.rst.txt
@@ -74,6 +74,11 @@ Develop mode
   pip install -r requirements.txt
   python setup.py develop
 
+New releases are frequent so always make sure that you have the latest version:
+
+.. code-block:: shell
+
+  pip install --upgrade pingouin
 
 Quick start
 ============
@@ -155,15 +160,17 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 4. Test the normality of the data
 #################################
 
-.. code-block:: ipython3
+.. code-block:: python
 
-   from pingouin import normality
+   from pingouin import normality, multivariate_normality
    # Return a boolean (true if normal) and the associated p-value
-   normality(x, y)
+   print(normality(x, y))                                 # Univariate normality
+   print(multivariate_normality(np.column_stack((x, y)))) # Multivariate normality
 
 .. parsed-literal::
 
-   [False,  True], [2.71e-04, 0.552]
+   (array([False,  True]), array([0., 0.552]))
+   (False, 0.00018)
 
 ------------
 
