### 2018-10-10 #999 @raphaelvallat
Hi @cMadan!
Just released version 0.1.10 of Pingouin, is it possible to update the review to this version?
Thank you - Raphael
