### 2018-10-11 #999 @cMadan
@raphaelvallat, that's not a problem. I've updated the version number in the initial comment.

@snastase @nhejazi @kellieotto, would you be available and interested to review this submission? Thanks for considering!
