+++ pingouin/__init__.py
@@ -10,5 +10,6 @@ from .correlation import *
 from .circular import *
 from .pairwise import *
 from .power import *
+from .regression import *
 
 __version__ = "0.1.10"
