### 2018-10-26 94ea787
Updated docs
