+++ pingouin/regression.py
@@ -1,9 +1,8 @@
 import numpy as np
-from scipy.stats import t
-# from scipy.stats import norm
+import pandas as pd
+from scipy.stats import t, norm
 
-
-__all__ = ['linear_regression']
+__all__ = ['linear_regression', 'mediation_analysis']
 
 
 def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
@@ -12,7 +11,7 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     Parameters
     ----------
     X : np.array or list
-        Predictor array. Shape = (n_samples, n_features) or (n_samples,).
+        Predictor(s). Shape = (n_samples, n_features) or (n_samples,).
     y : np.array or list
         Dependent variable. Shape = (n_samples).
     add_intercept : bool
@@ -53,6 +52,7 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     1. Simple linear regression
 
         >>> import numpy as np
+        >>> from pingouin import linear_regression
         >>> np.random.seed(123)
         >>> mean, cov, n = [4, 6], [[1, 0.5], [0.5, 1]], 30
         >>> x, y = np.random.multivariate_normal(mean, cov, n).T
@@ -123,11 +123,171 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     ll = coef - marg_error
     ul = coef + marg_error
 
-    if ll.sum() > ul.sum():
-        ll, ul = ul, ll
-
     # Create dict
     stats = {'coef': coef, 'se': beta_se, 'tvals': tvals, 'pvals': pvals,
              'rsquared': rsquared, 'adj_rsquared': adj_rsquared,
              'll': ll, 'ul': ul}
     return stats
+
+
+def _point_estimate(data, x, m, y, idx):
+    """Point estimate of indirect effect based on bootstrap sample."""
+    # Mediator model (M ~ X)
+    beta_m = linear_regression(data[x].iloc[idx], data[m].iloc[idx],
+                               add_intercept=True, coef_only=True)
+    # Full model (Y ~ X + M)
+    beta_y = linear_regression(data[[x, m]].iloc[idx], data[y].iloc[idx],
+                               add_intercept=True, coef_only=True)
+    # Point estimate
+    return beta_m[1] * beta_y[2]
+
+
+def _bias_corrected_interval(ab_estimates, sample_point, n_boot, alpha=0.05):
+    """Get (1 - alpha) * 100 bias-corrected confidence interval estimate
+
+    Parameters
+    ----------
+    ab_estimates : 1d array-like
+        Array with bootstrap estimates for each sample.
+    sample_point : float
+        Indirect effect point estimate based on full sample.
+    n_boot : int
+        Number of bootstrap samples
+    alpha : float
+        Alpha for confidence interval
+
+    Returns
+    -------
+    CI : 1d array-like
+        Lower limit and upper limit bias-corrected confidence interval
+        estimates.
+    """
+    # Bias of bootstrap estimates
+    z0 = norm.ppf(np.sum(ab_estimates < sample_point) / n_boot)
+
+    # Adjusted intervals
+    adjusted_ll = norm.cdf(2 * z0 + norm.ppf(alpha / 2)) * 100
+    adjusted_ul = norm.cdf(2 * z0 + norm.ppf(1 - alpha / 2)) * 100
+    ll = np.percentile(ab_estimates, q=adjusted_ll)
+    ul = np.percentile(ab_estimates, q=adjusted_ul)
+    return np.array([ll, ul])
+
+
+def mediation_analysis(data=None, x=None, m=None, y=None, alpha=0.05,
[…]
