+++ docs/build/html/_sources/generated/pingouin.mediation_analysis.rst.txt
@@ -1,0 +1,6 @@
+pingouin.mediation\_analysis
+============================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: mediation_analysis
