+++ docs/build/html/_sources/generated/pingouin.linear_regression.rst.txt
@@ -1,0 +1,6 @@
+pingouin.linear\_regression
+===========================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: linear_regression
