+++ MANIFEST.in
@@ -10,6 +10,7 @@ include pingouin/datasets/circular.csv
 include pingouin/datasets/cochran.csv
 include pingouin/datasets/datasets.csv
 include pingouin/datasets/icc.csv
+include pingouin/datasets/mediation.csv
 include pingouin/datasets/pairwise_corr.csv
 include pingouin/datasets/rm_anova.csv
 include pingouin/datasets/rm_corr.csv
