+++ setup.cfg
@@ -7,7 +7,7 @@ test=pytest
 #     slow: mark a test as slow.
 
 [flake8]
-ignore = N806, N803
+ignore = N806, N803, W605, W503
 exclude =
     .git,
     __pycache__,
