### 2018-10-26 0c1e8f1
Removed deprecationwarning external
