+++ setup.cfg
@@ -7,7 +7,9 @@ test=pytest
 #     slow: mark a test as slow.
 
 [flake8]
-ignore = N806, N803, W605, W503
+# W605 : bug when math equation in numpydoc
+W W503, W504 : line-break with math operator
+ignore = N806, N803, W503, W504, W605
 exclude =
     .git,
     __pycache__,
