### 2018-10-26 ac83ccb
Temporarily removed ancovan testing since statsmodels/patsy crashes for Travis and AppVeyor
