### 2018-10-27 00ccec9
Added logistic_regression function
