+++ pingouin/regression.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 from scipy.stats import t, norm
 
-__all__ = ['linear_regression', 'mediation_analysis']
+__all__ = ['linear_regression', 'logistic_regression', 'mediation_analysis']
 
 
 def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
@@ -27,7 +27,7 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     Returns
     -------
     stats : dict
-        Linear regression output::
+        Linear regression summary::
 
         'coef' : regression coefficients
         'se' : standard error of the estimate
@@ -35,12 +35,14 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
         'pvals' : p-values
         'rsquared' : coefficient of determination (R2)
         'adj_rsquared' : adjusted R2
+        'll' : lower confidence interval
+        'ul' : upper confidence interval
 
     Notes
     -----
     Results have been compared against the sklearn library.
 
-    The coefficient of determination (R2) is defined as:
+    The coefficient of determination (:math:`R^2`) is defined as:
 
     .. math:: R^2 = 1 - (\dfrac{SS_{resid}}{SS_{total}})
 
@@ -62,12 +64,12 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
 
     2. Multiple linear regression
 
-        >>> np.random.seed(123)
+        >>> np.random.seed(42)
         >>> z = np.random.normal(size=n)
         >>> X = np.column_stack((x, z))
         >>> lm = linear_regression(X, y)
         >>> print(lm['coef'])
-            [ 4.45443375  0.37561055 -0.22153521]
+            [4.54123324 0.36628301 0.17709451]
 
     3. Using a Pandas DataFrame
 
@@ -75,13 +77,13 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
         >>> df = pd.DataFrame({'x': x, 'y': y, 'z': z})
         >>> lm = linear_regression(df[['x', 'z']], df['y'])
         >>> print(lm['coef'])
-            [ 4.45443375  0.37561055 -0.22153521]
+            [4.54123324 0.36628301 0.17709451]
 
     4. No intercept and return coef only
 
         >>> linear_regression(df[['x', 'z']], df['y'], add_intercept=False,
         >>>                   coef_only=True)
-            array([ 1.41420086, -0.11302849])
+            array([ 1.40935593, -0.2916508 ])
     """
     X = np.asarray(X)
     y = np.asarray(y)
@@ -126,6 +128,131 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     # Create dict
     stats = {'coef': coef, 'se': beta_se, 'tvals': tvals, 'pvals': pvals,
              'rsquared': rsquared, 'adj_rsquared': adj_rsquared,
+             'll': ll, 'ul': ul}
+    return stats
+
+
+def logistic_regression(X, y, coef_only=False, alpha=0.05):
+    """Binary logistic regression.
+
+    Parameters
+    ----------
+    X : np.array or list
+        Predictor(s). Shape = (n_samples, n_features) or (n_samples,).
+    y : np.array or list
+        Dependent variable. Shape = (n_samples).
+        Must be binary.
+    coef_only : bool
+        If True, return only the regression coefficients.
+    alpha : float
+        Alpha value used for the confidence intervals.
+        CI = [alpha / 2 ; 1 - alpha / 2]
+
+    Returns
+    -------
+    stats : dict
[…]
