+++ README.rst
@@ -45,6 +45,8 @@ It provides easy-to-grasp functions for computing several statistical functions:
 
 8. Circular statistics
 
+9. Linear/logistic regression and mediation analysis
+
 Pingouin is designed for users who want **simple yet exhaustive statistical functions**.
 
 For example, the :code:`ttest_ind` function of SciPy returns only the T-value and the p-value. By contrast,
@@ -62,11 +64,16 @@ Installation
 Dependencies
 ------------
 
-Pingouin requires :
+The main dependencies of Pingouin are :
 
 * NumPy
 * SciPy
 * Pandas
+
+In addition, some functions require :
+
+* Statsmodels
+* Scikit-learn
 
 User installation
 -----------------
