+++ docs/changelog.rst
@@ -9,7 +9,8 @@ v0.2.0 (November 2018)
 **Regression**
 
 a. Added linear_regression function.
-b. Added mediation_analysis function.
+b. Added logistic_regression function.
+c. Added mediation_analysis function.
 
 
 v0.1.10 (October 2018)
