+++ docs/build/html/_sources/generated/pingouin.logistic_regression.rst.txt
@@ -1,0 +1,6 @@
+pingouin.logistic\_regression
+=============================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: logistic_regression
