### 2018-10-27 ddccba6
Mediation with binary mediator
