### 2018-10-28 349fdcf
Added doc and better code for regression.py
