+++ pingouin/regression.py
@@ -15,8 +15,8 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     y : np.array or list
         Dependent variable. Shape = (n_samples).
     add_intercept : bool
-        If False, assume that the data are already centered. If True, add an
-        intercept to the model. In this case, the first value in the
+        If False, assume that the data are already centered. If True, add a
+        constant term to the model. In this case, the first value in the
         output dict is the intercept of the model.
     coef_only : bool
         If True, return only the regression coefficients.
@@ -29,25 +29,54 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
     stats : dict
         Linear regression summary::
 
+        'names' : name of variable(s) in the model (e.g. x1, x2...)
         'coef' : regression coefficients
         'se' : standard error of the estimate
         'tvals' : T-values
         'pvals' : p-values
-        'rsquared' : coefficient of determination (R2)
-        'adj_rsquared' : adjusted R2
+        'r2' : coefficient of determination (R2)
+        'adj_r2' : adjusted R2
         'll' : lower confidence interval
         'ul' : upper confidence interval
 
     Notes
     -----
-    Results have been compared against the sklearn library.
+    The beta coefficients of the regression are estimated using the
+    np.linalg.lstsq function.
+
+    It is generally recommanded to include a constant term (intercept) to the
+    model to limit the bias and force the residual mean to equal zero.
+    Note that intercept coefficient and p-values are however rarely meaningful.
+
+    The standard error of the estimates is a measure of the accuracy of the
+    prediction defined as:
+
+    .. math:: se = \sqrt{MSE \cdot (X^TX)^{-1}}
+
+    where :math:`MSE` is the mean squared error,
+
+    .. math:: MSE = \dfrac{\sum{(true - pred)^2}}{n - p - 1}
+
+    :math:`p` is the total number of explanatory variables in the model
+    (excluding the intercept) and :math:`n` is the sample size.
+
+    Using the coefficients and the standard errors, the T-values can be
+    obtained:
+
+    .. math:: T = \dfrac{coef}{se}
+
+    and the p-values can then be approximated using a T-distribution
+    with :math:`n - p - 1` degrees of freedom.
 
     The coefficient of determination (:math:`R^2`) is defined as:
 
     .. math:: R^2 = 1 - (\dfrac{SS_{resid}}{SS_{total}})
 
-    Unlike most other scores, :math:`R^2` score may be negative (it need not
-    actually be the square of a quantity R).
+    The adjusted :math:`R^2` is defined as:
+
+    .. math:: \overline{R}^2 = 1 - (1 - R^2) \dfrac{n - 1}{n - p - 1}
+
+    Results have been compared against sklearn, statsmodels and JASP.
 
     Examples
     --------
@@ -71,6 +100,17 @@ def linear_regression(X, y, add_intercept=True, coef_only=False, alpha=0.05):
         >>> print(lm['coef'])
             [4.54123324 0.36628301 0.17709451]
 
+    2. Convert the output dictionnary to a pandas DataFrame
+
+        >>> import pandas as pd
+        >>> df_lm = pd.DataFrame.from_dict(linear_regression(X, y))
+        >>> # Round to 3 decimals
+        >>> df_lm = df_lm.round(3)
+        >>> # Print column names
+        >>> print(df_lm.keys())
[…]
