+++ pingouin/tests/test_regression.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-from numpy.testing import assert_almost_equal
+from numpy.testing import assert_almost_equal, assert_equal
 from pingouin.tests._tests_pingouin import _TestPingouin
 from pingouin.regression import (linear_regression, logistic_regression,
                                  mediation_analysis)
@@ -10,10 +10,6 @@ from scipy.stats import linregress
 from sklearn.linear_model import LinearRegression
 
 df = read_dataset('mediation')
-
-# For logistic regression test
-np.random.seed(123)
-df['Ybin'] = np.random.randint(0, 2, size=df.shape[0])
 
 
 class TestRegression(_TestPingouin):
@@ -25,26 +21,34 @@ class TestRegression(_TestPingouin):
         # Simple regression
         lm = linear_regression(df['X'], df['Y'])  # Pingouin
         sc = linregress(df['X'].values, df['Y'].values)  # SciPy
+        assert_equal(lm['names'], ['Intercept', 'X'])
         assert_almost_equal(lm['coef'][1], sc.slope)
         assert_almost_equal(lm['coef'][0], sc.intercept)
         assert_almost_equal(lm['se'][1], sc.stderr)
-        assert_almost_equal(lm['pvals'][1], sc.pvalue, decimal=5)
-        assert_almost_equal(np.sqrt(lm['rsquared']), sc.rvalue)
+        assert_almost_equal(lm['pvals'][1], sc.pvalue)
+        assert_almost_equal(np.sqrt(lm['r2']), sc.rvalue)
 
         # Multiple regression with intercept
         X = df[['X', 'M']].values
         y = df['Y'].values
         lm = linear_regression(X, y)  # Pingouin
         sk = LinearRegression(fit_intercept=True).fit(X, y)  # SkLearn
+        assert_equal(lm['names'], ['Intercept', 'x1', 'x2'])
         assert_almost_equal(lm['coef'][1:], sk.coef_)
         assert_almost_equal(lm['coef'][0], sk.intercept_)
-        assert_almost_equal(sk.score(X, y), lm['rsquared'])
+        assert_almost_equal(sk.score(X, y), lm['r2'])
+        # Compare values to JASP
+        assert_equal([.605, .110, .101], np.round(lm['se'], 3))
+        assert_equal([3.145, 0.361, 6.321], np.round(lm['tvals'], 3))
+        assert_equal([0.002, 0.719, 0.000], np.round(lm['pvals'], 3))
+        assert_equal([.703, -.178, .436], np.round(lm['ll'], 3))
+        assert_equal([3.106, .257, .835], np.round(lm['ul'], 3))
 
         # No intercept
         lm = linear_regression(X, y, add_intercept=False)
         sk = LinearRegression(fit_intercept=False).fit(X, y)
         assert_almost_equal(lm['coef'], sk.coef_)
-        assert_almost_equal(sk.score(X, y), lm['rsquared'])
+        assert_almost_equal(sk.score(X, y), lm['r2'])
 
         # Test other arguments
         linear_regression(df[['X', 'M']], df['Y'], coef_only=True)
@@ -56,28 +60,29 @@ class TestRegression(_TestPingouin):
 
         # Simple regression
         lom = logistic_regression(df['X'], df['Ybin'])  # Pingouin
-        # Compare to statsmodels
-        assert_almost_equal(lom['coef'], [-0.0761, -0.0208], decimal=2)
-        assert_almost_equal(lom['se'], [0.722, 0.116], decimal=2)
[…]
