### 2018-10-28 63aa769
Avoid duplicate ss_res in linear_regression
