### 2018-10-28 b92874e
Better documentation of one-way ANOVA
