### 2018-10-28 04dbf56
Added welch anova and pairwise_gameshowell post hoc tests
