+++ pingouin/pairwise.py
@@ -5,10 +5,11 @@ import pandas as pd
 from itertools import combinations
 from pingouin.parametric import anova
 from pingouin.multicomp import multicomp
-from pingouin.effsize import compute_effsize
+from pingouin.effsize import compute_effsize, convert_effsize
 from pingouin.utils import _remove_rm_na, _export_table, _check_dataframe
 
-__all__ = ["pairwise_ttests", "pairwise_tukey", "pairwise_corr"]
+__all__ = ["pairwise_ttests", "pairwise_tukey", "pairwise_gameshowell",
+           "pairwise_corr"]
 
 
 def _append_stats_dataframe(stats, x, y, xlabel, ylabel, effects, alpha,
@@ -308,8 +309,6 @@ def pairwise_tukey(dv=None, between=None, data=None, alpha=.05,
         >>> df = read_dataset('anova')
         >>> pairwise_tukey(dv='Pain threshold', between='Hair color', data=df)
     '''
-    from itertools import combinations
-    from pingouin.effsize import convert_effsize
     from pingouin.external.qsturng import psturng
 
     # First compute the ANOVA
@@ -359,6 +358,152 @@ def pairwise_tukey(dv=None, between=None, data=None, alpha=.05,
                          'efsize': np.round(ef, 3),
                          'eftype': effsize,
                          })
+    return stats
+
+
+def pairwise_gameshowell(dv=None, between=None, data=None, alpha=.05,
+                         tail='two-sided', effsize='hedges'):
+    '''Pairwise Games-Howell post-hoc tests.
+
+    Parameters
+    ----------
+    dv : string
+        Name of column containing the dependant variable.
+    between: string
+        Name of column containing the between factor.
+    data : pandas DataFrame
+        DataFrame
+    alpha : float
+        Significance level
+    tail : string
+        Indicates whether to return the 'two-sided' or 'one-sided' p-values
+    effsize : string or None
+        Effect size type. Available methods are ::
+
+        'none' : no effect size
+        'cohen' : Unbiased Cohen d
+        'hedges' : Hedges g
+        'glass': Glass delta
+        'eta-square' : Eta-square
+        'odds-ratio' : Odds ratio
+        'AUC' : Area Under the Curve
+
+    Returns
+    -------
+    stats : DataFrame
+        Stats summary ::
+
+        'A' : Name of first measurement
+        'B' : Name of second measurement
+        'mean(A)' : Mean of first measurement
+        'mean(B)' : Mean of second measurement
+        'diff' : Mean difference
+        'SE' : Standard error
+        'tail' : indicate whether the p-values are one-sided or two-sided
+        'T-val' : T-values
+        'df' : adjusted degrees of freedom
+        'pval' : Games-Howell corrected p-values
+        'efsize' : effect sizes
+        'eftype' : type of effect size
+
+    Notes
+    -----
+    Games-Howell is very similar to the Tukey HSD post-hoc test but is much
+    more robust to heterogeneity of variances. While the
+    Tukey-HSD post-hoc is optimal after a classic one-way ANOVA, the
+    Games-Howell is optimal after a Welch ANOVA.
+    Games-Howell is not valid for repeated measures ANOVA.
+
+    Compared to the Tukey-HSD test, the Games-Howell test uses different pooled
+    variances for each pair of variables instead of the same pooled variance.
+
+    The T-values are defined as:
+
+    .. math::
+
[…]
