+++ pingouin/parametric.py
@@ -5,7 +5,7 @@ import pandas as pd
 from pingouin import (_check_dataframe, _remove_rm_na, _remove_na,
                       _export_table, bayesfactor_ttest, epsilon, sphericity)
 
-__all__ = ["ttest", "rm_anova", "rm_anova2", "anova", "anova2",
+__all__ = ["ttest", "rm_anova", "rm_anova2", "anova", "anova2", "welch_anova",
            "mixed_anova", "ancova", "ancovan"]
 
 
@@ -608,7 +608,7 @@ def anova(dv=None, between=None, data=None, detailed=False,
         ANOVA summary ::
 
         'Source' : Factor names
-        'SS' :Sums of squares
+        'SS' : Sums of squares
         'DF' : Degrees of freedom
         'MS' : Mean squares
         'F' : F-values
@@ -620,12 +620,12 @@ def anova(dv=None, between=None, data=None, detailed=False,
     rm_anova : One-way repeated measures ANOVA
     rm_anova2 : Two-way repeated measures ANOVA
     mixed_anova : Two way mixed ANOVA
+    welch_anova : One-way Welch ANOVA
     kruskal : Non-parametric one-way ANOVA
-
 
     Notes
     -----
-    The ANOVA is very powerful when the groups are normally distributed
+    The classic ANOVA is very powerful when the groups are normally distributed
     and have equal variances. However, when the groups have unequal variances,
     it is best to use the Welch ANOVA (`welch_anova`) that better controls for
     type I error (Liu 2015). The homogeneity of variances can be measured with
@@ -792,7 +792,7 @@ def anova2(dv=None, between=None, data=None, export_filename=None):
         ANOVA summary ::
 
         'Source' : Factor names
-        'SS' :Sums of squares
+        'SS' : Sums of squares
         'DF' : Degrees of freedom
         'MS' : Mean squares
         'F' : F-values
@@ -895,6 +895,161 @@ def anova2(dv=None, between=None, data=None, export_filename=None):
 
     aov = aov.reindex(columns=col_order)
     aov.dropna(how='all', axis=1, inplace=True)
+    # Export to .csv
+    if export_filename is not None:
+        _export_table(aov, export_filename)
+    return aov
+
+
+def welch_anova(dv=None, between=None, data=None, export_filename=None):
+    """One-way Welch ANOVA.
+
+    Parameters
+    ----------
+    dv : string
+        Name of column containing the dependant variable.
+    between : string
+        Name of column containing the between factor.
+    data : pandas DataFrame
+        DataFrame
+    export_filename : string
+        Filename (without extension) for the output file.
+        If None, do not export the table.
+        By default, the file will be created in the current python console
+        directory. To change that, specify the filename with full path.
+
+    Returns
+    -------
+    aov : DataFrame
+        ANOVA summary ::
+
+        'Source' : Factor names
+        'SS' : Sums of squares
+        'DF' : Degrees of freedom
+        'MS' : Mean squares
+        'F' : F-values
+        'p-unc' : uncorrected p-values
+        'np2' : Partial eta-square effect sizes
+
+    See Also
+    --------
+    anova : One-way ANOVA
+    rm_anova : One-way repeated measures ANOVA
+    rm_anova2 : Two-way repeated measures ANOVA
+    mixed_anova : Two way mixed ANOVA
+    kruskal : Non-parametric one-way ANOVA
+
+    Notes
+    -----
+    The classic ANOVA is very powerful when the groups are normally distributed
[…]
