+++ docs/build/html/_sources/generated/pingouin.pairwise_gameshowell.rst.txt
@@ -1,0 +1,6 @@
+pingouin.pairwise\_gameshowell
+==============================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: pairwise_gameshowell

[…]
