### 2018-10-28 f8767f8
Added doc to pairwise_tukey
