+++ docs/conf.py
@@ -133,7 +133,8 @@ html_theme_options = {
     'navbar_pagenav': False,
     'bootstrap_version': "3",
     'navbar_links': [
-                     ("API", "api"),
+                     ("Functions", "api"),
+                     ("Guidelines", "guidelines"),
                      ("Examples", "examples"),
                      ("What's new", "changelog")
                      ],
