+++ docs/build/html/_sources/api.rst.txt
@@ -2,8 +2,8 @@
 
 .. currentmodule:: pingouin
 
-API reference
-=============
+Functions
+=========
 
 ANOVA and T-test
 ----------------
