### 2018-10-29 4da9689
Added correlation guideline
