### 2018-10-29 2cadb05
Updated changelog
