### 2018-10-29 107115f
Added references for function + TOC in doc
