+++ docs/build/html/_sources/api.rst.txt
@@ -5,6 +5,10 @@
 Functions
 =========
 
+.. contents:: Table of Contents
+   :depth: 2
+
+
 ANOVA and T-test
 ----------------
 
