+++ docs/build/html/_sources/changelog.rst.txt
@@ -3,6 +3,10 @@
 What's new
 ##########
 
+.. contents:: Table of Contents
+   :depth: 2
+
+
 v0.2.0 (November 2018)
 ----------------------
 
@@ -10,6 +14,11 @@ v0.2.0 (November 2018)
 
 a. Added Welch ANOVA
 b. Added Games-Howell post-hoc test for one-way ANOVA with unequal variances
+
+**Documentation**
+
+a. Added Guidelines section with flow charts
+b. Renamed API section to Functions
 
 **Regression**
 
