### 2018-10-29 f573e73
Improved doc pairwise_tukey
