### 2018-10-29 a1dcf17
Added import to readme, index and notebook
