### 2018-10-29 ad3c0ab
Added rationale for function ttest_power
