### 2018-10-29 948dd03
Added rationale to anova_power
