### 2018-10-29 2255ebe
Added rationale to Cochran test
