### 2018-10-30 b8e3f62
Improved sphericity test code and doc
