### 2018-10-30 68ed9f2
Added rationale to homoscedasticity
