### 2018-10-30 176f130
Improved doc / code Mauchly
