### 2018-10-30 aea9e36
Added rationale for normality test
