### 2018-10-30 559e441
Added rationale to T-test
