### 2018-10-30 ceb6643
Changed 0.2.0 to dev in changelog
