### 2018-10-30 f3a2e8e
Replaced \mathtt{df} by v
