### 2018-11-03 b7d463c
Added one-versus-all in pairwise_corr function
