### 2018-11-03 6c9d7f2
Advanced column indexing in pairwise_corr function
