+++ pingouin/pairwise.py
@@ -574,10 +574,16 @@ def pairwise_corr(data, columns=None, tail='two-sided', method='pearson',
     data : pandas DataFrame
         DataFrame
     columns : list or str
-        Names of columns in data containing the all the dependant variables.
-        If columns is None, compute the pairwise correlations on all numeric
-        columns. If columns has a single element, compute the correlations
-        between this element and all the other numeric columns.
+        Column names in data ::
+
+        '["a", "b", "c"]' : combination between columns a, b, and c
+        '["a"]' : product between a and all the other numeric columns
+        '[["a"], ["b", "c"]]' : product between ["a"] and ["b", "c"]
+        '[["a", "d"], ["b", "c"]]' : product between ["a", "d"] and ["b", "c"]
+
+        Note that if column is not specified, then the function will return the
+        pairwise correlation between the combination of all the numeric columns
+        in data. See the examples section for more details on this.
     tail : string
         Indicates whether to return the 'two-sided' or 'one-sided' p-values
     method : string
@@ -624,28 +630,15 @@ def pairwise_corr(data, columns=None, tail='two-sided', method='pearson',
 
     Notes
     -----
-    The Pearson correlation coefficient measures the linear relationship
-    between two datasets. Strictly speaking, Pearson's correlation requires
-    that each dataset be normally distributed. Correlations of -1 or +1 imply
-    an exact linear relationship.
+    Please refer to the `pingouin.corr()` function for a description of the
+    different method. NaN are automatically removed from datasets.
 
-    The Spearman correlation is a nonparametric measure of the monotonicity of
-    the relationship between two datasets. Unlike the Pearson correlation,
-    the Spearman correlation does not assume that both datasets are normally
-    distributed. Correlations of -1 or +1 imply an exact monotonic
-    relationship.
-
-    Kendall’s tau is a measure of the correspondence between two rankings.
-    Values close to 1 indicate strong agreement, values close to -1 indicate
-    strong disagreement.
-
-    The percentage bend correlation (Wilcox 1994) is a robust method that
-    protects against univariate outliers.
-
-    The Shepherd's pi correlation (Schwarzkopf et al. 2012) is a robust method
-    that returns the equivalent of the Spearman's rho after outliers removal.
-
-    Please note that NaN are automatically removed from datasets.
+    This function is more flexible and gives a much more detailed
+    output than the `pandas.DataFrame.corr()` method (i.e. p-values,
+    confidence interval, Bayes Factor..). This comes however at
+    an increased computational cost. While this should not be discernible for
+    dataframe with less than 10,000 rows and/or less than 20 columns, this
+    function can be extremely slow for very large dataset.
 
     Examples
     --------
@@ -660,25 +653,21 @@ def pairwise_corr(data, columns=None, tail='two-sided', method='pearson',
 
     2. Robust two-sided correlation with uncorrected p-values
 
-        >>> from pingouin.datasets import read_dataset
-        >>> from pingouin import pairwise_corr, print_table
-        >>> data = read_dataset('pairwise_corr').iloc[:, 1:]
         >>> pairwise_corr(data, columns=['Openness', 'Extraversion',
         >>>                              'Neuroticism'], method='percbend')
 
     3. Export the results to a .csv file
 
-        >>> from pingouin.datasets import read_dataset
-        >>> from pingouin import pairwise_corr
-        >>> data = read_dataset('pairwise_corr').iloc[:, 1:]
         >>> pairwise_corr(data, export_filename='pairwise_corr.csv')
 
-    4. One-versus-all pairwise correlations
+    4. One-versus-others pairwise correlations
 
-        >>> from pingouin.datasets import read_dataset
-        >>> from pingouin import pairwise_corr
[…]
