### 2018-11-05 596134d
Added Gitter + Joss badge
