### 2018-11-05 d10fff8
Fix pytest-sugar crash travis
