+++ docs/build/html/_sources/index.rst.txt
@@ -15,15 +15,13 @@
 .. image:: https://codecov.io/gh/raphaelvallat/pingouin/branch/master/graph/badge.svg
     :target: https://codecov.io/gh/raphaelvallat/pingouin
 
-.. image:: https://mybinder.org/badge.svg
-    :target: https://mybinder.org/v2/gh/raphaelvallat/pingouin/master
-
 .. image:: http://joss.theoj.org/papers/d2254e6d8e8478da192148e4cfbe4244/status.svg
     :target: http://joss.theoj.org/papers/d2254e6d8e8478da192148e4cfbe4244
 
 .. image:: https://badges.gitter.im/owner/repo.png
     :target: https://gitter.im/pingouin-stats/Lobby
 
+----------------
 
 .. figure::  /pictures/logo_pingouin.png
   :align:   center
@@ -100,6 +98,8 @@ Chat
 
 If you have questions, please ask them in the public `Gitter chat <https://gitter.im/pingouin-stats/Lobby>`_
 
+.. image:: https://badges.gitter.im/owner/repo.png
+    :target: https://gitter.im/pingouin-stats/Lobby
 
 Quick start
 ============
