### 2018-11-05 #1026 @raphaelvallat
Hi @cMadan and @snastase!

I was hoping to release the version 0.2.0 of Pingouin by early next week and ideally I would really love to include the JOSS paper in it. Do you think that it would be possible to do the review by then? If not, no worries, I'll just include it in a future release.

Thank you and have a great day!
Raphael
