### 2018-11-05 03ed1d5
Regression dataframe as default output + T instead of T-val
