### 2018-11-06 fb46fea
Added n to corr and pairwise_corr
