+++ README.rst
@@ -157,11 +157,11 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 .. table:: Output
    :widths: auto
 
-   =====  ===========  =====  ========  =======  ======
-       r  CI95%           r2    adj_r2    p-val    BF10
-   =====  ===========  =====  ========  =======  ======
-   0.595  [0.3  0.79]  0.354     0.306    0.001  54.222
-   =====  ===========  =====  ========  =======  ======
+  ====  =====  ===========  =====  ========  =======  ======
+     n      r  CI95%           r2    adj_r2    p-val    BF10
+  ====  =====  ===========  =====  ========  =======  ======
+    30  0.595  [0.3  0.79]  0.354     0.306    0.001  54.222
+  ====  =====  ===========  =====  ========  =======  ======
 
 ------------
 
@@ -178,11 +178,11 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 .. table:: Output
    :widths: auto
 
-   =====  ===========  =====  ========  =======
-       r  CI95%           r2    adj_r2    p-val
-   =====  ===========  =====  ========  =======
-   0.561  [0.25 0.77]  0.315     0.264    0.002
-   =====  ===========  =====  ========  =======
+   ====  =====  ===========  =====  ========  =======
+      n      r  CI95%           r2    adj_r2    p-val
+   ====  =====  ===========  =====  ========  =======
+     30  0.561  [0.25 0.77]  0.315     0.264    0.002
+   ====  =====  ===========  =====  ========  =======
 
 ------------
 
@@ -326,14 +326,13 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 .. table:: Output
   :widths: auto
 
-  ===  ===  ========  =========  =====  =============  =====  ========  =====  =======  ======
-  X    Y    method    tail           r  CI95%             r2    adj_r2      z    p-unc    BF10
-  ===  ===  ========  =========  =====  =============  =====  ========  =====  =======  ======
-  X    Y    pearson   two-sided  0.707  [0.14 0.92]    0.500     0.357  0.881    0.022   3.227
-  X    Z    pearson   two-sided  0.283  [-0.42  0.77]  0.080    -0.183  0.291    0.428   0.321
-  Y    Z    pearson   two-sided  0.105  [-0.56  0.69]  0.011    -0.271  0.105    0.772   0.243
-  ===  ===  ========  =========  =====  =============  =====  ========  =====  =======  ======
-
+  ===  ===  ========  =========  ===  =====  =============  =====  ========  =====  =======  ======
+  X    Y    method    tail         n      r  CI95%             r2    adj_r2      z    p-unc    BF10
+  ===  ===  ========  =========  ===  =====  =============  =====  ========  =====  =======  ======
+  X    Y    pearson   two-sided   10  0.707  [0.14 0.92]    0.500     0.357  0.881    0.022   3.227
+  X    Z    pearson   two-sided   10  0.283  [-0.42  0.77]  0.080    -0.183  0.291    0.428   0.321
+  Y    Z    pearson   two-sided   10  0.105  [-0.56  0.69]  0.011    -0.271  0.105    0.772   0.243
+  ===  ===  ========  =========  ===  =====  =============  =====  ========  =====  =======  ======
 
 10. Convert between effect sizes
 ################################

[…]
