### 2018-11-06 af89f31
width auto in readme
