+++ README.rst
@@ -157,11 +157,11 @@ Try before you buy! Click on the link below and navigate to the notebooks folder
 .. table:: Output
    :widths: auto
 
-  ====  =====  ===========  =====  ========  =======  ======
-     n      r  CI95%           r2    adj_r2    p-val    BF10
-  ====  =====  ===========  =====  ========  =======  ======
-    30  0.595  [0.3  0.79]  0.354     0.306    0.001  54.222
-  ====  =====  ===========  =====  ========  =======  ======
+   ====  =====  ===========  =====  ========  =======  ======
+      n      r  CI95%           r2    adj_r2    p-val    BF10
+   ====  =====  ===========  =====  ========  =======  ======
+     30  0.595  [0.3  0.79]  0.354     0.306    0.001  54.222
+   ====  =====  ===========  =====  ========  =======  ======
 
 ------------
 
