### 2018-11-06 81f30fb
Fixed bug BF10 pairwise_corr if data.shape > 1000
