+++ pingouin/tests/test_pairwise.py
@@ -121,3 +121,7 @@ class TestPairwise(_TestPingouin):
         pairwise_corr(data, columns=[['Age', 'IQ'], []])
         pairwise_corr(data, columns=['Age', 'Gender', 'IQ', 'Wrong'])
         pairwise_corr(data, columns=['Age', 'Gender', 'Wrong'])
+        # Test with more than 1000 columns (BF10 not computed)
+        data1500 = pd.concat([data, data, data], ignore_index=True)
+        pcor1500 = pairwise_corr(data1500, method='pearson')
+        assert 'BF10' not in pcor1500.keys()
