### 2018-11-06 69201e0
Removed useless variable + added numpy>=1.15 requirements
