### 2018-11-10 00dc0bc
Improved code pairwise_ttests
