### 2018-11-10 e5176b2
Started working on better pairwise_ttests
