+++ pingouin/pairwise.py
@@ -12,8 +12,8 @@ __all__ = ["pairwise_ttests", "pairwise_tukey", "pairwise_gameshowell",
            "pairwise_corr"]
 
 
-def _append_stats_dataframe(stats, x, y, xlabel, ylabel, effects, alpha,
-                            paired, df_ttest, ef, eftype, time=np.nan):
+def _append_stats_dataframe(stats, x, y, xlabel, ylabel, alpha, paired,
+                            df_ttest, ef, eftype, time=np.nan):
     stats = stats.append({
         'A': xlabel,
         'B': ylabel,
@@ -22,7 +22,6 @@ def _append_stats_dataframe(stats, x, y, xlabel, ylabel, effects, alpha,
         # Use ddof=1 for unibiased estimator (pandas default)
         'std(A)': np.round(x.std(ddof=1), 3),
         'std(B)': np.round(y.std(ddof=1), 3),
-        'Type': effects,
         'Paired': paired,
         'tail': df_ttest.loc['T-test', 'tail'],
         # 'Alpha': alpha,
@@ -31,7 +30,7 @@ def _append_stats_dataframe(stats, x, y, xlabel, ylabel, effects, alpha,
         'BF10': df_ttest.loc['T-test', 'BF10'],
         'efsize': ef,
         'eftype': eftype,
-        'Time': time}, ignore_index=True)
+        'Time': time}, ignore_index=True, sort=False)
     return stats
 
 
@@ -120,9 +119,9 @@ def pairwise_ttests(dv=None, between=None, within=None, subject=None,
     from pingouin.parametric import ttest
 
     # Safety checks
+    effects_orig = '%s' % effects  # For later use (ensure deep copy)
     effects = 'within' if between is None else effects
     effects = 'between' if within is None else effects
-
     _check_dataframe(dv=dv, between=between, within=within, subject=subject,
                      effects=effects, data=data)
 
@@ -132,19 +131,42 @@ def pairwise_ttests(dv=None, between=None, within=None, subject=None,
     if not isinstance(alpha, float):
         raise ValueError('Alpha must be float')
 
-    # Remove NAN in repeated measurements
-    if within is not None and data[dv].isnull().values.any():
-        data = _remove_rm_na(dv=dv, within=within, subject=subject, data=data)
+    # Check if we have multiple between or within factors
+    multiple_between = False
+    multiple_within = False
+    if isinstance(between, list):
+        if len(between) > 1:
+            multiple_between = True
+        else:
+            between = between[0]
+
+    if isinstance(within, list):
+        if len(within) > 1:
+            multiple_within = True
+        else:
+            within = within[0]
+
+    if multiple_within is True and multiple_between is True:
+        raise ValueError("Multiple between and within factors are",
+                         "currently not supported. Please select only one.")
+
+    multiple = any([multiple_within, multiple_between])
+    effects = 'all' if multiple is True else effects
 
     # Initialize empty variables
     stats = pd.DataFrame([])
     ddic = {}
 
-    # OPTION A: simple main effects
+    # OPTION A: SIMPLE MAIN EFFECTS, WITHIN OR BETWEEN
     if effects.lower() in ['within', 'between']:
         # Compute T-tests
         paired = True if effects == 'within' else False
         col = within if effects == 'within' else between
+
+        # Remove NAN in repeated measurements
+        if within is not None and data[dv].isnull().values.any():
+            data = _remove_rm_na(dv=dv, within=within, subject=subject,
+                                 data=data)
 
         # Extract effects
         labels = data[col].unique().tolist()
@@ -154,7 +176,6 @@ def pairwise_ttests(dv=None, between=None, within=None, subject=None,
         # Number and labels of possible comparisons
         if len(labels) >= 2:
[…]
