### 2018-11-10 b151585
MAJOR UPDATE PAIRWISE_TTESTS FUNCTION
