### 2018-11-12 #1026 @snastase
@cMadan @raphaelvallat Sorry for holding this up!

I’m not a very experienced developer (also my first JOSS review), so take this as the perspective of intermediate-skill-level user. Overall, I think this is a great contribution. Documentation and examples are stellar. I work through the checklist and point out some minor issues below (most of which are installation issues that may be peculiar to my environment).

General checklist: general checks are all good (version has progressed to 0.1.10).

Functional checklist: Installation with pip (on Mac) was smooth, although it might be worth including a note for folks like me who still have a python 2.7 installation floating around; e.g., `pip3 install pingouin`. I get some incompatibility warnings for requirements (`mock`, `future`, `jsonschema`, `decorator`), but these do not derail installation. Install via pip was smooth enough on a Linux machine. Module imports fine in IPython. Developer install via git clone was smooth.

I followed the “10 Minutes to Pingouin” walkthrough in IPython on both Mac and Linux machines, and most of it ran smoothly. I ran into the following handful of errors.

- When I run post-hoc tests using `pairwise_ttests` on my Linux machine, I get a `TypeError: drop() got an unexpected keyword argument 'columns'` at line 243 in `pairwise.py`... not sure why. The `stats` object is a pandas DataFrame and I’m evidently on pandas version 0.20.3 , so that’s probably the issue. However this error did not occur with the Mac installation.

- On my Mac machine, I was unable to `from pingouin import linear_regression` probably due to a pandas incompatibility (v.0.23.1). Upgrade to pandas v.0.23.4 didn’t seem to fix this issue. On my Linux installation I had a similar issue when importing `linear_regression` which returned `ModuleNotFoundError: No module named 'pandas.compat'`. Same issue when importing `mediation_analysis`. I just noticed in the Quick Start notebook that maybe these features are provisional and planned for release in v.0.2.0…?

I also walked through all of the tutorial notebooks on Binder. This is an excellent approach to interactive examples. I ran into the following issues:

- In the 00_QuickStart notebook running on Binder, the `‘mixed_anova’` dataset does not seem to exist in `pingouin.datasets`.

- In the 04_Correlations notebook, in the second code cell of the “Advanced column selection” where you do `pairwise_corr(df, columns='Age')`, I get a pandas error: `KeyError: 'A'`. The following cells in that section do not run correctly either.

Documentation checklist: Regarding the statement of need, it might be helpful to include in the introductory documentation how this relates to adjacent packages such as `statsmodels`, whether it provides complementary functionality, etc. I spot-checked the tests, and they seem comprehensive enough. Community guidelines are fine, e.g,. code of conduct, gitter, etc. Documentation (tutorials, docstrings) and examples are _excellent_. The flowchart guidelines are really nice too.

Paper checklist: Everything in order. You might do a final pass over the references to make sure formatting is correct; e.g., most of the journal names in your list are abbreviated, but the Berens reference is not—should be “J. Stat. Softw.”; also capitalize the proceedings name for the McKinney reference.

Random highly-specific feature request: In the `corr` function, it would be nice to have separate implementations of Kendall’s tau-a, tau-b, and tau-c. This occurs to me because in representational similarity analysis in neuroimaging, the recommended metric for comparing representational similarity matrices is Kendall’s tau-a (Nili et al., PLOS Comp Biol, 2014). As far as I know, there’s no python implementation of tau-a.

Again, sorry for the delay! Let me know if I can be of help.

Sam
