### 2018-11-12 a8fd50b
Prepare switch to 0.2.0
