+++ JOSS/paper.bib
@@ -97,7 +97,7 @@
 
 
 @ARTICLE{Berens2009,
-  title    = "{CircStat}: A {MATLAB} Toolbox for Circular Statistics",
+  title    = "CircStat: A {MATLAB} Toolbox for Circular Statistics",
   author   = "Berens, Philipp",
   abstract = "Directional data is ubiquitious in science. Due to its circular
               nature such data cannot be analyzed with commonly used
@@ -114,18 +114,18 @@
               how to apply them to data. Finally, we analyze a dataset from
               neurophysiology to demonstrate the capabilities of the CircStat
               toolbox.",
-  journal  = "Journal of Statistical Software, Articles",
+  journal  = "J. Stat. Softw.",
   volume   =  31,
   number   =  10,
   pages    = "1--21",
   year     =  2009
 }
 
-@InProceedings{ Pandas,
-  author    = { Wes McKinney },
-  title     = { Data Structures for Statistical Computing in Python },
-  booktitle = { Proceedings of the 9th Python in Science Conference },
-  pages     = { 51 - 56 },
-  year      = { 2010 },
-  editor    = { Stefan van der Walt and Jarrod Millman }
+@InProceedings{Pandas,
+  author    = "Wes McKinney",
+  title     = "Data Structures for Statistical Computing in Python",
+  booktitle = "Proceedings of the 9th Python in Science Conference",
+  pages     = "51--56",
+  year      = 2010,
+  editor    = "Stefan van der Walt and Jarrod Millman"
 }
