+++ pingouin/__init__.py
@@ -12,4 +12,4 @@ from .pairwise import *
 from .power import *
 from .regression import *
 
-__version__ = "0.1.10"
+__version__ = "0.2.0"
