+++ docs/changelog.rst
@@ -7,8 +7,8 @@ What's new
    :depth: 2
 
 
-v0.2.0 (dev)
-------------
+v0.2.0 (November 2018)
+----------------------
 
 **ANOVA**
 
