+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy>=1.15
-scipy
-pandas
+scipy>=1.1
+pandas>=0.23
