### 2018-11-12 a6374d6
Added FAQ section to docs
