### 2018-11-12 2662f4b
Added several questions to FAQ
