+++ docs/build/html/_sources/faq.rst.txt
@@ -3,20 +3,22 @@
 FAQ
 ***
 
+Python
+------
+
 .. ----------------------------- INTRO -----------------------------
 .. raw:: html
 
     <div class="panel-group">
       <div class="panel panel-default">
         <div class="panel-heading">
-          <h4 class="panel-title">
-            <a data-toggle="collapse" href="#collapse_python">How can I start with Python and Pingouin?</a>
-          </h4>
+          <h5 class="panel-title">
+            <a data-toggle="collapse" href="#collapse_python">I am new to Python, how can I install Python and Pingouin on my computer?</a>
+          </h5>
         </div>
         <div id="collapse_python" class="panel-collapse collapse">
           <div class="panel-body">
 
-If you are completely new to Python, this section is for you.
 To install Python  on your computer, you should use `Anaconda <https://conda.io/docs/index.html>`_, a Python distribution which natively includes all the most important packages. Then, open the newly installed Anaconda prompt and type:
 
 .. code-block:: bash
@@ -56,9 +58,9 @@ Now, let's do a simple paired T-test using Pingouin:
 
     <div class="panel panel-default">
       <div class="panel-heading">
-        <h4 class="panel-title">
+        <h5 class="panel-title">
           <a data-toggle="collapse" href="#collapse_import">How to import and use Pingouin?</a>
-        </h4>
+        </h5>
       </div>
       <div id="collapse_import" class="panel-collapse collapse">
         <div class="panel-body">
@@ -66,15 +68,16 @@ Now, let's do a simple paired T-test using Pingouin:
 .. code-block:: python
 
     # 1) Import the full package
+    # --> Best if you are planning to use several Pingouin functions.
     import pingouin as pg
     pg.ttest(x, y)
 
     # 2) Import specific functions
+    # --> Best if you are planning to use only this specific function.
     from pingouin import ttest
     ttest(x, y)
 
-
-.. ----------------------------- DATA FORMAT -----------------------------
+.. ----------------------------- STATSMODELS -----------------------------
 .. raw:: html
 
           </div>
@@ -83,14 +86,71 @@ Now, let's do a simple paired T-test using Pingouin:
 
     <div class="panel panel-default">
       <div class="panel-heading">
-        <h4 class="panel-title">
-          <a data-toggle="collapse" href="#collapse_data">I want to analyze a spreadsheet (.csv or .xlsx file), how should I format my data?</a>
-        </h4>
+        <h5 class="panel-title">
+          <a data-toggle="collapse" href="#collapse_sm">What are the differences between statsmodels and Pingouin?</a>
+        </h5>
+      </div>
+      <div id="collapse_sm" class="panel-collapse collapse">
+        <div class="panel-body">
+
+`Statsmodels <https://www.statsmodels.org/stable/index.html>`_ is a great statistical Python package that provides several advanced functions (regression, GLM, time-series analysis) as well as an R-like syntax for fitting models. However, statsmodels can be quite hard to grasp and use for Python beginners and/or users who just want to perform simple statistical tests. The goal of Pingouin is not to replace statsmodels but rather to provide some easy-to-use functions to perform the most widely-used statistical tests. The two packages are therefore complementary, for instance one can use Pingouin first to do some exploratory data analysis in a few lines of code and then switch to statsmodels if a more exhaustive output is needed.
+
+.. ----------------------------- SCIPY -----------------------------
+.. raw:: html
+
+          </div>
+        </div>
+      </div>
+
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h5 class="panel-title">
+          <a data-toggle="collapse" href="#collapse_scp">What are the differences between scipy.stats and Pingouin?</a>
+        </h5>
+      </div>
[…]
