### 2018-11-13 #1026 @raphaelvallat
Hi @snastase!

Thank you so much for your review and your positive feedback, I really appreciate it!

Please find below the answers to the main points that you mentioned in your review:

1) I never tried to run Pingouin on Python 2.7. My idea when I started working on the project was that it would be a Python 3 only package. I uploaded the installation instructions to emphasize that it works better with Python 3 and that users should be careful when using Python 2.7

2) Thank you for noticing the `pandas.drop()` error in the `pairwise_ttests()` function. I realized that the `columns` argument has only been added since Pandas 0.21. I have therefore updated the requirements.txt file to make sure that the pip installation also updates the pandas version.

3) I have just uploaded the version 0.2.0 of Pingouin to PyPi, which should solve all the bugs that you have encountered in the Binder notebooks (i.e. no linear regression, missing example datasets...).
Just make sure to upgrade your version: `pip install --upgrade pingouin`

4) I liked the idea of adding a section explaining how Pingouin relates to similar packages. Therefore, I have just added a [FAQ](https://raphaelvallat.github.io/pingouin/build/html/faq.html) section in the documentation to explain the differences between Statsmodels, Scipy and Pingouin, as well as to answer some of the most basic functions that users may ask when they first start to use Python and/or Pingouin.

5) I have updated the Berens reference  in the paper, however, I am not able to capitalize the proceedings name for the McKinney reference (the automatic PDF generation somehow does not take it into account).

6) I was not aware of these alternatives metrics of the Kendall tau. The Pingouin implementation relies on the `scipy.stats.kendalltau()` function, which, according to the documentation, uses the tau-b when ties are present and tau-a in absence of ties. Therefore, it seems that only the tau-c implementation is missing. I will definitely keep that in mind and maybe try to add it in a future release.

Thank you again, Samuel, for your review! Please let me know if there is anything else that I can do.
Have a great day,
Raphael
