### 2018-11-14 bb212fd
Added documentation parametric ci to compute_esci
