### 2018-11-14 3d73bd4
MAJOR updates in bootstrap and parametric CI
