+++ pingouin/effsize.py
@@ -2,43 +2,32 @@
 # Date: April 2018
 import numpy as np
 from pingouin.utils import _check_eftype, _remove_na
-from pingouin.distribution import homoscedasticity
+# from pingouin.distribution import homoscedasticity
 
 
-__all__ = ["compute_esci", "convert_effsize", "compute_effsize",
-           "compute_effsize_from_t"]
+__all__ = ["compute_esci", "compute_boot_esci", "convert_effsize",
+           "compute_effsize", "compute_effsize_from_t"]
 
 
-def compute_esci(x=None, y=None, ef=None, nx=None, ny=None, alpha=.95,
-                 method='parametric', n_boot=2000, eftype='cohen',
-                 return_dist=False):
-    """Bootstrapped or parametric confidence intervals of an effect size.
+def compute_esci(stat=None, nx=None, ny=None, eftype='d', confidence=.95,
+                 decimals=2):
+    """Parametric confidence intervals around an effect size or a
+    correlation coefficient.
 
     Parameters
     ----------
-    x, y : int
-        Data vectors (required for bootstrapping only)
-    ef : float
+    stat : float
         Original effect size. Must be either a correlation coefficient or a
         Cohen-type effect size (Cohen d or Hedges g).
-        Required for parametric method only.
     nx, ny : int
         Length of vector x and y.
-    alpha : float, optional
-        Confidence interval (0.95 = 95%)
-    method : string
-        Computation method
-        Available methods are ::
-
-        'parametric' : uses standard deviation of effect sizes.
-        'bootstrap' : uses a bootstrapping procedure (pivotal CI).
-    n_boot : int
-        Number of permutations for the bootstrap procedure
     eftype : string
-        Effect size type for the bootstrapping procedure.
-    return_dist : boolean
-        If True, return the distribution of permutations (e.g. useful for a
-        posteriori plotting)
+        Effect size type. Must be 'r' (correlation) or 'cohen'
+        (Cohen d or Hedges g).
+    confidence : float
+        Confidence level (0.95 = 95%)
+    decimals : int
+        Number of rounded decimals.
 
     Returns
     -------
@@ -106,110 +95,210 @@ def compute_esci(x=None, y=None, ef=None, nx=None, ny=None, alpha=.95,
 
     Examples
     --------
-    1. Compute the 95% **parametric** confidence interval of an effect size
-       given the two sample sizes.
+    1. Confidence interval of a Pearson correlation coefficient
+
+        >>> import pingouin as pg
+        >>> x = [3, 4, 6, 7, 5, 6, 7, 3, 5, 4, 2]
+        >>> y = [4, 6, 6, 7, 6, 5, 5, 2, 3, 4, 1]
+        >>> nx, ny = len(x), len(y)
+        >>> stat = np.corrcoef(x, y)[0][1]
+        >>> ci = pg.compute_esci(stat=stat, nx=nx, ny=ny, eftype='r')
+        >>> print(stat, ci)
+            0.7468280049029223 [0.27 0.93]
+
+    2. Confidence interval of a Cohen d
+
+        >>> import pingouin as pg
+        >>> x = [3, 4, 6, 7, 5, 6, 7, 3, 5, 4, 2]
+        >>> y = [4, 6, 6, 7, 6, 5, 5, 2, 3, 4, 1]
+        >>> nx, ny = len(x), len(y)
+        >>> stat = pg.compute_effsize(x, y, eftype='cohen')
+        >>> ci = pg.compute_esci(stat=stat, nx=nx, ny=ny, eftype='cohen')
+        >>> print(stat, ci)
+            0.1537753990658328 [-0.68  0.99]
[…]
