+++ docs/generated/pingouin.compute_boot_esci.rst
@@ -1,0 +1,6 @@
+pingouin.compute\_boot\_esci
+============================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: compute_boot_esci
