### 2018-11-15 73a405a
Renamed function compute_bootci and added univariate function
