+++ pingouin/effsize.py
@@ -5,13 +5,13 @@ from pingouin.utils import _check_eftype, _remove_na
 # from pingouin.distribution import homoscedasticity
 
 
-__all__ = ["compute_esci", "compute_boot_esci", "convert_effsize",
+__all__ = ["compute_esci", "compute_bootci", "convert_effsize",
            "compute_effsize", "compute_effsize_from_t"]
 
 
-def compute_esci(stat=None, nx=None, ny=None, eftype='d', confidence=.95,
+def compute_esci(stat=None, nx=None, ny=None, eftype='cohen', confidence=.95,
                  decimals=2):
-    """Parametric confidence intervals around an effect size or a
+    """Parametric confidence intervals around a Cohen d or a
     correlation coefficient.
 
     Parameters
@@ -143,22 +143,28 @@ def compute_esci(stat=None, nx=None, ny=None, eftype='d', confidence=.95,
     return np.round(ci, decimals)
 
 
-def compute_boot_esci(x, y, func='pearson', method='cper', paired=False,
-                      confidence=.95, n_boot=2000, decimals=2, seed=None):
-    """Bootstrapped confidence intervals.
+def compute_bootci(x, y=None, func='pearson', method='cper', paired=False,
+                   confidence=.95, n_boot=2000, decimals=2, seed=None,
+                   return_dist=False):
+    """Bootstrapped confidence intervals of univariate and bivariate functions.
 
     Parameters
     ----------
-    x, y : 1D-arrays  or lists
-        Samples.
+    x : 1D-array or list
+        First sample. Required for both bivariate and univariate functions.
+    y : 1D-array, list, or None
+        Second sample. Required only for bivariate functions.
     func : str or custom function
         Function to compute the bootstrapped statistic.
         Accepted string values are::
 
-        'pearson': Pearson correlation coefficient
-        'spearman': Spearman correlation coefficient
-        'cohen': Cohen d effect size
-        'hedges': Hedges g effect size
+        'pearson': Pearson correlation (bivariate, requires x and y)
+        'spearman': Spearman correlation (bivariate)
+        'cohen': Cohen d effect size (bivariate)
+        'hedges': Hedges g effect size (bivariate)
+        'mean': Mean (univariate, requires only x)
+        'std': Standard deviation (univariate)
+        'var': Variance (univariate)
     method : str
         Method to compute the confidence intervals::
 
@@ -167,15 +173,18 @@ def compute_boot_esci(x, y, func='pearson', method='cper', paired=False,
         'cper': Bias corrected percentile method (default)
     paired : boolean
         Indicates whether x and y are paired or not. Only useful when computing
-        Cohen d or Hedges g bootstrapped confidence intervals.
+        bivariate Cohen d or Hedges g bootstrapped confidence intervals.
     confidence : float
         Confidence level (0.95 = 95%)
     n_boot : int
-        Number of bootstrap iterations.
+        Number of bootstrap iterations. The higher, the better, the slower.
     decimals : int
         Number of rounded decimals.
     seed : int or None
         Random seed for generating bootstrap samples.
+    return_dist : boolean
+        If True, return the confidence intervals and the bootstrapped
+        distribution  (e.g. for plotting purposes).
 
     Returns
     -------
[…]
