+++ docs/build/html/_sources/generated/pingouin.compute_boot_esci.rst.txt
@@ -1,6 +1,0 @@
-pingouin.compute\_boot\_esci
-============================
-
-.. currentmodule:: pingouin
-
-.. autofunction:: compute_boot_esci
