+++ docs/changelog.rst
@@ -12,7 +12,7 @@ v0.2.1 (dev)
 
 **Effect size**
 
-a. Separated compute_esci and compute_boot_esci
+a. Separated compute_esci and compute_bootci
 b. Added corrected percentile method and normal approximation to bootstrap
 c. Fixed bootstrapping method
 
