+++ docs/generated/pingouin.compute_bootci.rst
@@ -1,0 +1,6 @@
+pingouin.compute\_bootci
+========================
+
+.. currentmodule:: pingouin
+
+.. autofunction:: compute_bootci
