### 2018-11-15 c726e22
Added outlier + method to skipped internal function
