### 2018-11-16 454ea0c
updated setup.py
