### 2018-11-16 b484842
Test ancovan with new version of patsy
