### 2018-11-17 #1026 @raphaelvallat
Hi @snastase and @cMadan !

I am just wondering what should be the next step(s) for this review. Is there anything else that I can do?
This is my very first JOSS paper so I am new to this process,

Thank you and have an excellent day!
Raphael
