### 2018-11-17 #1026 @cMadan
@raphaelvallat, I think you're all set for now.

@snastase, can you look over the reviewer checklist in the first post in this issue and check off what's all sorted, and see what still needs to be addressed? Apart from that, please look over @raphaelvallat's response and revisions and let us know what you think of the current state of the submission.
