### 2018-11-19 #1026 @snastase
Okay, I think the author has adequately addressed my comments! New v0.2.0 fixes my import problems for `linear_regression` and `mediation_analysis`. The `pairwise_ttests` functionality now works normally for me. I like the mention of statsmodels in the FAQ. I've completed the review checklist above and am satisfied with the current state of the package—I endorse it for acceptance/publication.
