### 2018-11-19 #1026 @cMadan
@snastase, perfect, thank you for your thorough review!

@raphaelvallat, you're almost done! I just need the DOI for an archived version of the current code (i.e., deposit it on Zenodo or FigShare).
