### 2018-11-19 #1026 @raphaelvallat
@snastase Great! thank you so much, Samuel, for your review!

@cMadan The DOI number of the current release is: https://zenodo.org/badge/latestdoi/127584791

[![DOI](https://zenodo.org/badge/127584791.svg)](https://zenodo.org/badge/latestdoi/127584791)

Have a great day!
Raphael
